+++ files/constants.py
@@ -1,0 +1,6 @@
+import mimetypes
+
+SUPPORTED_IMAGES_TYPES = (
+    mimetypes.types_map[".jpg"],
+    mimetypes.types_map[".png"],
+)
