+++ files/views.py
@@ -4,7 +4,7 @@ from rest_framework import permissions, status
 from rest_framework.generics import get_object_or_404
 from rest_framework.response import Response
 
-from files.helpers import FileAPI
+from files.service import FileAPI
 from files.models import UserFile
 from files.serializers import UserFileSerializer
